Add files via upload
These files are used to generate an html file that contains a map of the song sparrow locations recorded. It is specifically designed to work with my data. You would need to modify the code to make it work for other CSVs. My data used UTM coordinates so there is a section in the code that addresses this and converts them to Lat, Lon. Including my generated map file here as well so you can see what it should look like.